project_neo4j/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
py2neo==2021.2.3
py2neo==4.3.0
redis==4.5.5
